% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addWeights.R
\name{addWeights}
\alias{addWeights}
\title{Create and Initialize the weights of the network
\code{addWeights}}
\usage{
addWeights(NETall = NETall, SdWinit = 2, MeanWinit = 1,
  Scaling = T, Adam = T, LSTM = T)
}
\arguments{
\item{NETall}{df, the AMoNet network architecture}

\item{SdWinit}{numeric, standard deviation of the random initialization, using the \code{rnorm} base function}

\item{MeanWinit}{numeric, mean of the random initialization, using the \code{rnorm} base function}

\item{Scaling}{bool, scales weights, Xavier initialization}

\item{Adam}{bool, create weigths for advanced optimizers such as Adam, including and used also for Momentum or RMSprop}

\item{LSTM}{bool, is the unit LSTM}
}
\value{
an updated AMoNet with weights
}
\description{
Create and Initialize the weights of the network
\code{addWeights}
}
