% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDoc.R
\docType{data}
\name{Boundaries}
\alias{Boundaries}
\title{Boundaries}
\description{
Boundaries are limits of the hyper-parameter search.
}
\details{
\code{print(Boundaries)} to visualize default hyper-parameters.
}
