% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NETforOptS3.R
\name{build.AMoNet}
\alias{build.AMoNet}
\title{Build the AMoNet network}
\usage{
build.AMoNet(object, MeanWinit = Default$MeanWinit,
  SdWinit = Default$SdWinit, RestrictedBuilding = F,
  nblayers = Default$nblayers, MinConnect = Default$MinConnect,
  FilterCGS = F, LSTM = Default$LSTM, Optimizer = Default$Optimizer,
  Interval = Default$Interval, InteractionBase = OMNI,
  Phenotypes = NULL, KeepPhenotypes = F, MECA = names_MECA,
  RestrictedBase = F, NameProj = "My AMoNet",
  no_cores = Default$no_cores)
}
\arguments{
\item{object}{*AMoNet* object. Has to be intiated from queries with default \code{AMoNet()} function.}

\item{MeanWinit}{numeric. The mean value of the normal distribution used to initiate the weights.}

\item{SdWinit}{numeric. The standard deviation of the normal distribution used to initiate the weights.}

\item{RestrictedBuilding}{boolean. During the network buidling phase, restrict new species only to feedback connections within the net.}

\item{nblayers}{integer. Number of layers from the gene queries to explore within the InteractionBase (number of search steps). Should be >=1.}

\item{MinConnect}{integer. During the filtering phase, what is the minimal value of connections to keep a node. If MinConnect=1, you'll discard all nodes that are only "passengers".}

\item{FilterCGS}{boolean. Do you want to keep cancer genes only}

\item{LSTM}{boolean. Will you use LSTM unit update in the learning phase?}

\item{Optimizer}{boolean. Will you use another optimizer than GD (ie Adam, Momentum, RMSprop), to initiate related weights.}

\item{Interval}{integer or NULL. Number of outputs to the network. To simulate descrete time survival, use Interval to set the number of time intervals. For label dimensions of 1, set \code{Interval=1}. Outputs will be connected to phenotypes. Default is 10.}

\item{InteractionBase}{data frame. The iteraction base used to build the network. Should be a data.frame in a sif-like format, with at least 3 colmuns: source_hgnc, interaction_directed_signed, target_hgnc}

\item{Phenotypes}{data frame of the same format as InteractionBase but with species as source_hgnc and names of phenotypes as target_hgnc. Can be a table of gene sets for example. Default loaded are Hallmarks and immune gene sets.}

\item{KeepPhenotypes}{boolean. Phenotypes are required to connect output(s)}

\item{MECA}{character vector or NULL. the phenotypes you want to keep in the network.}

\item{RestrictedBase}{boolean. Restrict the building of the nets to genes related to the phenotypes selected with MECA.}

\item{NameProj}{character. The name of your project.}
}
\value{
An object of class *AMoNet*. Stores the name of the project, the network, the initial states \code{iStates=NULL} and \code{Ct=NULL} if LSTM.
}
\description{
Build the AMoNet network
}
\details{
Run \code{print(Default)} to check and optionnaly set default values.
}
