% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDoc.R
\docType{data}
\name{CGS}
\alias{CGS}
\title{CGS}
\description{
Cancer Gene Census from Cosmic database.
}
\details{
Used in the \code{build()} function if \code{FilterCGS=TRUE} to reduce *AMoNet* network buiding to cancer specific genes.
Used in the \code{LoadCleanTCGA()} function if number of \code{Species} selected is > 800. It selects only CGS genes among user-selected \code{Species}. Motivated by problems in downloading more \code{Species}.
}
\references{
\url{https://cancer.sanger.ac.uk/census}
}
