% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{CommandArgs}
\alias{CommandArgs}
\title{retreive the hyperparameters to search from the commandline and write it on globalenv()}
\usage{
CommandArgs()
}
\description{
retreive the hyperparameters to search from the commandline and write it on globalenv()
}
