% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDoc.R
\docType{data}
\name{Default}
\alias{Default}
\title{Default}
\description{
Default the default hyper-parameters used for building, simulation and training of *AMoNet* object.
}
\details{
\code{print(Default)} to visualize default hyper-parameters.
}
