% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/essAlone.R
\name{ess}
\alias{ess}
\title{ess: multi start simulations}
\usage{
ess(net, Quant = 30, Init, MUT, iStates, y, treatmt, no_cores = 3)
}
\arguments{
\item{net}{*AMoNet* object}

\item{Quant}{numeric. Number of evaluations}

\item{Init}{matrix. Used to set some of the initial states (iStates). To force some initial states during learning set adaptive_iStates=F. Init=NULL otherwise.}

\item{MUT}{list. Corresponds to mutations with names and corresponding values (list of vectors). If MUT=NULL, a wild type *AMoNet* is simulated.. Patients ordered the same as y.}

\item{iStates}{matrix. Initial states of the simulations.}

\item{y}{numeric or matrix. labels}

\item{treatmt}{list. The same for than newMUT, with treatments' targets and corresponding values (list of vectors).}

\item{no_cores}{numeric. If Parallel=TRUE, set the number of cores to parallelize on. Default is 4. Can detect and set to available no_cores if inferior to user defined no_cores.}

\item{treatmt}{list. The same for than newMUT, with treatments' targets and corresponding values (list of vectors).}
}
\description{
ess: multi start simulations
}
