% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDoc.R
\docType{data}
\name{GenesSelectHall}
\alias{GenesSelectHall}
\title{GenesSelectHall}
\description{
GenesSelectHall are cancer hallmarks related gene sets from MSigDB hallmarks.
}
\references{
\url{http://software.broadinstitute.org/gsea/msigdb/}
}
