% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyperP.R
\name{HyperP}
\alias{HyperP}
\title{Randomise hyper parameters}
\usage{
HyperP(C = C, Default = Default, Boundaries = Boundaries)
}
\arguments{
\item{C}{characeter vector. Name(s) of the hyper-parameters to optimize}

\item{Default}{list. Default hyper-parameters from \code{Default}}

\item{Boundaries}{list. Default hyper-parameters Boundaries from \code{Boundaries}}
}
\description{
Randomise hyper parameters
}
