% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLoadingFunc.R
\name{LoadCleanTCGA}
\alias{LoadCleanTCGA}
\title{Loads cBioportal data}
\usage{
LoadCleanTCGA(object, Species = NULL, Param = "", RestrictUnique = T,
  PLOT = F, FunctionalAnnot = T, organ = Default$organ)
}
\arguments{
\item{object}{AMoNet object.}

\item{Species}{character vectors. Gene names in Hugo nomenclature.}

\item{Param}{character vectors. Parameters to randomly search for.}

\item{RestrictUnique}{boolean. In case of multiple patients with same molecular features, these are removed.}

\item{PLOT}{boolean. Should the \code{LoadCleanTCGA} function return a waterfall plot of the mutational landscape?
Default is \code{FALSE}}

\item{FunctionalAnnot}{boolean. Should the mutations be binary (1 or 0, if \code{FunctionalAnnot=FALSE}) or quantitative (if \code{FunctionalAnnot=TRUE}).
Default is \code{TRUE}.}

\item{organ}{regexp or unique character. Data will be loaded for the type of organs selected. \code{NULL} is accepted and will select default value, \code{Default$organ="lung"}.}
}
\value{
A list with data downloaded from cBioportal, and corresponding to Default parameters.
y is a descrete time survival matrix whose number of intervals depend on the number of output of the model.
SurvData is the classical survival data of 2 dimensions with rowns: Time, Status and patients in colmuns.
MUTa is a mutation data matrix, either binary if FunctionalAnnot=F or numeric is FunctionalAnnot=T
MUT is a list format of MUTa generated with the \code{MutMatToList} function
Init is a normalized expression matrix
}
\description{
cBioportal stores publically available TCGA data.
}
\details{
requires cgdsr package
requires internet connection
}
