% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutMatToList.R
\name{MutMatToList}
\alias{MutMatToList}
\title{Computes a list of mutations per sample}
\usage{
MutMatToList(MUTa = MUTa)
}
\arguments{
\item{MUTa}{matrix. Mutation matrix with samples in rows, genes in columns, and numeric values between 0 and 1 for functional impact}
}
\value{
A list of altereed genes, per sample, with a vector numeric values for each patient/gene with signs corresponding to either gain or loss of function
}
\description{
Computes a list of mutations per sample
}
