% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetBuilding.R
\name{NetBuilding}
\alias{NetBuilding}
\title{This function builds a network and clean it from genes' query and interaction database}
\usage{
NetBuilding(GENESman = GENESman, treatmt = treatmt, OMNI = OMNI,
  nblayers = 2, FamilyGene = NULL, WRITE = F, MinConnect = 1,
  VeryRestricted = F, no_cores = 4)
}
\arguments{
\item{GENESman}{character vector. A vector of gene(s) selected to build the *AMoNet* object.}

\item{treatmt}{character vector. A vector of gene(s) targeted by treatment(s). Used in to build the *AMoNet* object.}

\item{OMNI}{data frame. Interaction database with columns ordered as: source_hgnc, interaction_directed_signed, target_hgnc corresponding to source species, interation type (ACTIVATE or INHIBIT), target species, respectively.}

\item{nblayers}{integer. Number of layers from the gene queries to explore within the InteractionBase (OMNI by default), i.e number of search steps. Should be >=1.}

\item{FamilyGene}{data frame. Restrict the building of the nets to genes related to the phenotypes selected with MECA. Default are \code{GenesSelectImmuno} and/or \code{GenesSelecHall}}

\item{WRITE}{boolean. Should the network be saved as csv?}

\item{MinConnect}{integer. During the filtering phase, what is the minimal value of connections to keep a node. If MinConnect=1, you'll discard all nodes that are only "passengers".}

\item{VeryRestricted}{boolean. A restriction during the search in the PPI database to keep only the new nodes that have already a partner in the network at the previous time step.}

\item{no_cores}{numeric. Set the number of cores to parallelize on. Package parallel required.}
}
\description{
This function builds a network and clean it from genes' query and interaction database
}
