% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NETforOpt.R
\name{NETforOpt}
\alias{NETforOpt}
\title{This function do the first steps of building of the net (for future optimization)}
\usage{
NETforOpt(WRITE = F, NETall = NULL, GENESman = c("PDCD1", "CD274",
  "CTLA4", "KRAS", "EGFR", "TP53"), treatmt = c("PDCD1", "CD274",
  "CTLA4"), InteractionBase = OMNI, MeanWinit = 0, SdWinit = 0.1,
  RestrictedBuilding = F, nblayers = 1, MinConnect = 1,
  FilterCGS = F, LSTM = F, Adam = F, Activity = T,
  addOutputs = NULL, Phenotypes = NULL, KeepPhenotypes = F,
  MECA = NULL, RestrictedBase = F, NameProj = "Model",
  no_cores = no_cores)
}
\arguments{
\item{MECA}{: Character vector or NULL. the phenotypes you want to keep in the network.}

\item{NETall:}{data frame or NULL. to build a netwrok or your network table.}

\item{GENESman:}{Character vector. The genes query, in symbols corresponding to the InteractionBase.}

\item{treatmt:}{Character vector. The genes targets of your treatment, in symbols corresponding to the InteractionBase.}

\item{InteractionBase:}{data frame. The iteraction base used to build the network. Should be a data.frame in a sif-like format, with at least 3 colmuns: source_hgnc, interaction_directed_signed, target_hgnc}

\item{MeanWinit:}{Numeric. The mean value of the normal distribution used to initiate the weights.}

\item{SdWinit:}{Numeric. The standard deviation of the normal distribution used to initiate the weights.}

\item{RestrictedBuilding:}{Boolean. During the network buidling phase, restrict new species only to feedback connections within the net.}

\item{nblayers:}{Integer. Number of layers from the gene queries to explore within the InteractionBase (number of search steps). Should be >=1.}

\item{MinConnect:}{Integer. During the filtering phase, what is the minimal value of connections to keep a node. If MinConnect=1, you'll discard all nodes that are only "passengers".}

\item{FilterCGS:}{Boolean. Do you want to keep cancer genes only}

\item{LSTM:}{Boolean. Will you use LSTM unit update in the learning phase?}

\item{Adam:}{Boolean. Will you use another optimizer than GD (ie Adam, Momentum, RMSprop), to initiate related weights.}

\item{Activity:}{Boolean. Run simulationn of the network to return activity values for each species.}

\item{addOutputs:}{Integer or NULL. Do you need outputs to the net, for example to simulate survival per interval. Specify the number of outputs. Outputs will be connected to phenotypes.}

\item{Phenotypes:}{data frame of the same format as InteractionBase but with species as source_hgnc and names of phenotypes as target_hgnc. Can be a table of gene sets for example. Default loaded are Hallmarks and immune gene sets.}

\item{KeepPhenotypes:}{Boolean. Phenotypes are required to connect output(s)}

\item{RestrictedBase:}{Boolean. Restrict the building of the nets to genes related to the phenotypes selected with MECA}

\item{NameProj:}{Character. The name of your project}
}
\description{
This function do the first steps of building of the net (for future optimization)
}
