% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDoc.R
\docType{data}
\name{OMNI}
\alias{OMNI}
\title{OMNI}
\description{
OMNI is a directed protein protein interaction (PPI) database from omnypath pooled with and immune models (in house).
}
\references{
\url{http://omnipathdb.org/} \url{https://acsn.curie.fr/ACSN2/ACSN2.html}
}
\keyword{data}
