% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3functions.R
\name{plot.AMoNet.history}
\alias{plot.AMoNet.history}
\alias{plot}
\title{Plot history of training performed with the \code{AMoNet()} function}
\usage{
plot(x, ...)
}
\arguments{
\item{x}{of class *AMoNet.history*. Can be foudn in *AMoNet* object \code{$history}}

\item{...}{additional optional arguments.}
}
\description{
Plot history of training performed with the \code{AMoNet()} function
}
