% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotAMoNet.R
\name{plot.AMoNet}
\alias{plot.AMoNet}
\title{Plot *AMoNet* objects}
\usage{
\method{plot}{AMoNet}(x = net, network = F, simulations = F,
  LEGEND = F, PDF = F, Npat = NULL, DIR = file.path(getwd(),
  "tmp"), Species = NULL, ...)
}
\arguments{
\item{network}{boolean. Whether to plot the *AMoNet* network.}

\item{simulations}{boolean. Whether to plot the results of *AMoNet* simulations.}

\item{LEGEND}{Boolean. If \code{TRUE}, add a legend to the plot}

\item{PDF}{Boolean. If \code{TRUE}, save a pdf version to the DIR argument}

\item{Npat}{numeric vector. The sample(s) for which simulations are ploted.}

\item{DIR}{path. If \code{PDF=TRUE}, set the path to save pdf. Default is the \code{/tmp/} directory of AMoNet package.}

\item{Species}{character. The name of gene(s) for which simulations are ploted.}

\item{net}{*AMoNet* object, S3 class}
}
\value{
In case the model is optimized it allows returning the updated Boolean relations.
}
\description{
Can either plot the network with layers and reclassify Boolean interactions based on learned parameters (if \code{network=TRUE})
or plot simulations performed (if \code{simulation=TRUE})
}
