% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3functions.R
\name{plot.predict.AMoNet}
\alias{plot.predict.AMoNet}
\alias{plot}
\title{Plot predictions generated from the \code{predicte()} function}
\usage{
plot(x=net$Predict, ...)
}
\arguments{
\item{x}{object of class *predict.AMoNet* generated with the \code{predict()} function and accessible with *AMoNet* object subobject \code{object$Predict}.}

\item{...}{additional arguments for the \code{plot()} function.}
}
\value{
The relation between true and predicted values.
}
\description{
Plot predictions generated from the \code{predicte()} function
}
