% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunTCGAopt.R
\name{PlotAndPredict}
\alias{PlotAndPredict}
\title{Plots and predictions within AMoNet grid search pipeline}
\usage{
PlotAndPredict(net, DIR = file.path(getwd(), "tmp"))
}
\arguments{
\item{net}{*AMoNet* object.}

\item{DIR}{path to file to plot. Default in tmp/}
}
\value{
Metriccs for training and validation
}
\description{
Plots and predictions within AMoNet grid search pipeline
}
