% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreBuiltNetS3.R
\name{prebuilt.AMoNet}
\alias{prebuilt.AMoNet}
\title{Use already built network in AMoNet}
\usage{
prebuilt.AMoNet(object, NETall = NULL, Weights = NULL,
  MeanWinit = Default$MeanWinit, SdWinit = Default$SdWinit,
  LSTM = Default$LSTM, Optimizer = Default$Optimizer,
  Outputs = "Output", NameProj = "My AMoNet")
}
\arguments{
\item{object}{*AMoNet* object.}

\item{NETall}{data frame. The newtork you provide to *AMoNet*. See the "details" section for format requirements.}

\item{Weights}{numeric vector. User-defined weights for each interactions. Same lengths as number of rows in NETall. B terms are initialized to zeros. Leave comment if other options needed.}

\item{MeanWinit}{numeric. The mean value of the normal distribution used to initiate the weights.}

\item{SdWinit}{numeric. The standard deviation of the normal distribution used to initiate the weights.}

\item{LSTM}{boolean. Will you use LSTM unit update in the learning phase?}

\item{Optimizer}{boolean. Will you use another optimizer than GD (ie Adam, Momentum, RMSprop), to initiate related weights.}

\item{Outputs}{character vector. The name(s) of the outputs of the network.}

\item{NameProj}{character. The name of your project.}
}
\value{
An object of class *AMoNet*. Stores the name of the project, the network, the initial states \code{iStates=NULL} and \code{Ct=NULL} if LSTM.
}
\description{
Use already built network in AMoNet
}
\details{
The network files provided in the NETall argument should have the following format:
* It should be a R data.frame object
* First column name is "source_hgnc" and stores the names of source species in character format.
* Second column name is "interaction_directed_signed" and encode the type of interaction for each row with the following characters: "ACTIVATE",  "INHIBIT".
* Third column name is "target_hgnc" and stores the names of target species in character format.
* Rownames are not used.

Example of the first 3 rows of a data frame supported:

source_hgnc interaction_directed_signed target_hgnc
SRC                     INHIBIT          PTEN
PTEN                    ACTIVATE         E2F1
CSNK2A1                 INHIBIT          PTEN

Names of genes are in Hugo nomenclature. It is necessary to use data from TCGA with the \code{LoadCleanTCGA()} function, for exeample.
For any other purpose, the nomenclaure is at the discretion of the user.

Run \code{print(Default)} to check and optionnaly set default values.
}
