% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3functions.R
\name{predict.AMoNet}
\alias{predict.AMoNet}
\alias{predict}
\title{Perform predictions from a trained *AMoNet* object}
\usage{
predict(object, newy=NULL, newInit=NULL, newiStates=NULL, newMUT=NULL, newtreatmt=NULL, Logic = Default$Logic, Mode = Default$Mode, MinSteps = Default$MinStepsForward, LSTM = Default$LSTM, Parallelize = Default$Parallelize, no_cores = Default$no_cores, ValMut = Default$ValMut)
}
\arguments{
\item{object}{*AMoNet* object}

\item{newy}{numeric or matrix. new labels used to compute metrics}

\item{newInit}{matrix. Used to set some of the initial states (iStates). Init=NULL otherwise.}

\item{newiStates}{matrix. Initial states of the simulations.}

\item{newMUT}{list. Corresponds to mutations with names and corresponding values (list of vectors). If MUT=NULL, wild type *AMoNet* is simulated. Patients ordered the same as newy.}

\item{newtreatmt}{list. In the same for than newMUT, with treatments' targets and corresponding values (list of vectors).}

\item{SplitType}{character. With split set to use between: "Train", "Val" or NULL}

\item{ValMut}{numeric. Multicplicative factor for the effect of mutations. Default to 50.}
}
\value{
*AMoNet* object with \code{$Predict} of class *predict.AMoNet* and results of simulations in \code{$TotAttractors}
}
\description{
Perform predictions from a trained *AMoNet* object
}
\details{
The \code{predict()} runs new simulations of the *AMoNet* object with new data.
Can be used to evaluate the performance of the model on a validation set.
}
