% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3functions.R
\name{print.AMoNet.history}
\alias{print.AMoNet.history}
\title{Print training cost}
\usage{
\method{print}{AMoNet.history}(x = net$history, ...)
}
\arguments{
\item{x}{of class "AMoNet.history" created in the \code{$history} of the *AMoNet* object.}
}
\description{
Print training cost
}
