% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunTCGAopt.R
\name{RunTCGAopt}
\alias{RunTCGAopt}
\title{Wrapper function to run the AMonet workflow, including hyper-parameters grid search, on TCGA data}
\usage{
RunTCGAopt(Param = c("nblayers", "MinConnect"),
  DIR = file.path(getwd(), "model"), NameProj = "HallmarksLungHN",
  GENESman = c("EGFR", "KRAS", "TP53", "MTOR"), treatmt = NULL,
  SelectMECA = "HALLMARK", organ = "luad", eSS = F, NewNet = T,
  no_cores = 3, KeepData = T, PartitionSplit = 0.7,
  Default = Default, Boundaries = Boundaries)
}
\arguments{
\item{Param}{character vector: the hyper-parameter(s) to randomly test whithin this run.}

\item{DIR}{directory: directory where the models are stored}

\item{NameProj}{character: name of your project. If at least one model with the name of your project NameProj is in the DIR directory, the script will load it and use it.}

\item{GENESman}{character vector: gene(s)' name(s) (in Hugo nomenclature) used to build the AMoNET network. If "all", all genes related to SelectMECA (gene sets) will be selected. If "frequent", 95\% most frequently altered genes from the cancer gene census (CGS) list in the cohort will be selected.}

\item{treatmt}{character vector: gene(s)' name(s) targeted by simulated treatment(s). Added to GENEman toused to build the AMoNET network.}

\item{SelectMECA}{character in regexp format: selection of one or several gene sets related to biological mecanisms. Names gene sets are available with the command \code{print(names_MECA)}.}

\item{organ}{character in regexp format: selection of one or several organs or cancer types. Names of organs are available with the command \code{print(names_MECA)}}

\item{eSS}{boolean to perform multistart or not. Default is set to \code{FALSE}. Recommandation is to use \code{eSS=TRUE} within a hyper-parameters search and in case of \code{Param = "MeanWinit"} and/or \code{Param = "SdWinit"} eSS \code{TRUE}.}

\item{NewNet}{boolean: when set to FALSE, if at least one model with the name of your project NameProj is in the DIR directory, the script will load it and use it. If set to TRUE, a new molecular network will be built using arguments GENESman and default parameters (Interval, MinConnect, nblayers).}
}
\value{
todo AMoNet object
}
\description{
AMonet workflow comprises building, training and simulating a network model to predict survival of patients from TCGA genomics WES data.
}
\details{
This function can be run:
1) Alone to train an AMoNet model with fixed hyper-parameters are either user-defined or the default one are used. For default hyper-parameters values call \code{print(Default)}.
2) In array and iteratively from command line together with the \code{PlotAndPredict} function to perform hyper-parameters search. Hyperparameter selected in the \code{Param} argument are thus randomly selected in batches and best results selected.

Requirements:
Uses the pre-configured /model and /tmp paths in AMoNet package directories (checks todo)

To set manually any hyper-paramters, change it into the \code{print(Default)} object.
}
