% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectGridSearch.R
\name{SelectGridSearch}
\alias{SelectGridSearch}
\title{Select best AMoNet model and update hyper-parameters from grid-search}
\usage{
SelectGridSearch(NameProjbase = "LungRegular", DIR = file.path(getwd(),
  "/model"), Default = Default, Bondaries = Bondaries, ValSelect = T)
}
\arguments{
\item{NameProjbase}{character. Name of the project}

\item{DIR}{path to the models}

\item{Default}{list. Default hyper-parameter values to update with best model's hyper-parameters (or predicted one if sufficient confidence)}

\item{Bondaries}{list. Boundaries of hyper-parameter seach to update with the confidences of predicted best hyper-parameters}

\item{ValSelect}{boolean. Whether the metrics for validation should be used for model selection. Default is \code{TRUE}}
}
\value{
Best AMoNet object with udpated hyper-parameters and their search space
}
\description{
Select best AMoNet model and update hyper-parameters from grid-search
}
