% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectGTExOptNet.R
\name{SelectGTExOptNet}
\alias{SelectGTExOptNet}
\title{Title}
\usage{
SelectGTExOptNet(NameProjbase = "LungRegular", DIR = file.path(getwd(),
  "/model"), NETall = NETall, Default = Default,
  Adam = !is.null(Default$Optimizer), LSTM = Default$LSTM,
  Bondaries = Bondaries, ValSelect = T)
}
\arguments{
\item{ValSelect}{boolean. Whether the metrics for validation should be used for model selection. Default is \code{TRUE}}
}
\description{
Title
}
