% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3functions.R
\name{simulate.AMoNet}
\alias{simulate.AMoNet}
\alias{simulate}
\title{Simulates *AMoNet*}
\usage{
simulate(object, nsim=Default$MinStepsForward, seed=NULL, Init=NULL, iStates=NULL, MUT=NULL, treatmt=NULL, Logic=Default$Logic, Mode=Default$Mode, ValMut=Default$ValMut, LSTM=Default$LSTM, Ct=NULL, Parallel=Default$Parallelize, no_cores=Default$no_cores, Steepness=1)
}
\arguments{
\item{object}{*AMoNet* object, S3 class.}

\item{nsim}{numeric. Number of simulation steps. Default to 5. Ideally set the number of simulation steps to achieve stable states of *AMoNet* activities.}

\item{seed}{generic parameter. Not used in *AMoNet* simulation as it doesn't involve stochastic behavior.}

\item{Init}{matrix. Used to set some of the initial states (iStates). To force some initial states during learning set adaptive_iStates=F. Init=NULL otherwise.}

\item{iStates}{matrix. Initial states can be either a vector, data.frame, matrix or list of vectors, with name of all species as columns and example as rows}

\item{MUT}{list. Corresponds to mutations with names and corresponding values (list of vectors). If MUT=NULL, a wild type *AMoNet* is simulated.. Patients ordered the same as y.}

\item{treatmt}{list. The same for than newMUT, with treatments' targets and corresponding values (list of vectors).}

\item{Logic}{character. The update function used for the simulations. Choices are: "Sigmoid" (default), "Boolean","tanh" or "ReLU".}

\item{Mode}{character. The type of simulation to perform. Choice are "LAYER" (layer based, default), "ASYNC" (assynchronous update), "SYNC" (synchronous update)}

\item{ValMut}{numeric. Multicplicative factor for the effect of mutations. Default to 50.}

\item{LSTM}{boolean. Should a Long-Short Term Memory (LSTM) unit architecture be used?}

\item{Ct}{matrix. If LSTM=TRUE, Ct is used to set the initial states of the correponding values in LSTM units.}

\item{Parallel}{boolean. Should the simulation run in parallel (TRUE by default)? Not recommended for simulations of less than 10 examples.}

\item{no_cores}{numeric. If Parallel=TRUE, set the number of cores to parallelize on. Default is 4. Can detect and set to available no_cores if inferior to user defined no_cores. to}

\item{Steepness}{numeric. Can be used to modulate the stepness of the sigmoid curve (default to 1).}
}
\value{
*AMoNet* object with \code{$TotAttractors}, a 4-dimensions array with: D1 the number of examples, D2 the results of the linear function (z=ax+b), D3 the results of the activator function (defined by the \code{Logic} argument) and D4 the number of iterations.
}
\description{
Simulates *AMoNet*
}
\details{
The S3 \code{simulate()} function calls the core \code{BoolSimul()} function. Simulations are vectorized and optinally paralellized.
The ASYNC mode perform a fully assynchroneous simulation that may be relatively long. The LAYER mode is an innovation in *AMoNet* to further speed up the simulations and nevertheless optain similar activities to the ASYNC mode. SYNC mode is not recommended.
}
