% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3functions.R
\name{write.AMoNet.network}
\alias{write.AMoNet.network}
\title{Write the network of the *AMoNet* object}
\usage{
write.AMoNet.network(object, file = stop("'file' must be specified"))
}
\arguments{
\item{object}{*AMoNet* object, S3 class.}

\item{file}{a (writable binary-mode) connection or the name of the file where the data will be saved.}
}
\description{
Write the network of the *AMoNet* object
}
\details{
Useful to export a .csv file for visualization and other applications
}
