% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunTCGAopt.R
\name{PlotAndPredict}
\alias{PlotAndPredict}
\title{Plots and predictions within AMoNet grid search pipeline}
\usage{
PlotAndPredict(net, DIR = file.path(getwd(), "tmp/new.pdf"))
}
\arguments{
\item{net}{*AMoNet* object.}

\item{DIR}{path and name to plot pdf.}
}
\value{
Metrics for training and validation in \code{$predict_[]} from *AMoNet* object
}
\description{
Plots and predictions within AMoNet grid search pipeline
}
