% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunTCGAopt.R
\name{PlotAndPredict}
\alias{PlotAndPredict}
\title{Plots and predictions within AMoNet grid search pipeline}
\usage{
PlotAndPredict(net)
}
\arguments{
\item{net}{*AMoNet* object, after training and with split defined (net$TrainSplit)}
}
\value{
Metrics for training and validation in \code{$predict_[]} from *AMoNet* object
Plots of training procedures and predictions in the tmp/ directory
}
\description{
Plots and predictions within AMoNet grid search pipeline
}
