% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3functions.R
\name{RandomiStates}
\alias{RandomiStates}
\title{Generates random initial states}
\usage{
RandomiStates(object, RETURN=F)
}
\arguments{
\item{object}{*AMoNet* object, S3 class.}
}
\value{
iStates, a dataframe that contains initial states
}
\description{
Used prior to \code{simulate()} or \code{AMoNet()} with available \code{$Data$y} in the *AMoNet* object
}
\examples{
\dontrun{RandomiStates(object)}
}
